Added extra check for new lib patch to check if the original files exist.

+++ scripts/patch_new_lib.py
@@ -16,13 +16,17 @@ def patch_new_lib():
     custom_new_cpp = os.path.join(env['PROJECT_DIR'], "patch", "new", "src", "new.cpp")
     custom_new_h = os.path.join(env['PROJECT_DIR'], "patch", "new", "src", "new.h")
 
-    # Check if the custom files exist, and if so, replace the framework's files
-    if os.path.exists(custom_new_cpp) and os.path.exists(custom_new_h):
-      print(f"Replacing {new_cpp_path} with custom new.cpp")
-      shutil.copy(custom_new_cpp, new_cpp_path)
-      print(f"Replacing {new_h_path} with custom new.h")
-      shutil.copy(custom_new_h, new_h_path)
+    # Check if the framework's original files exist before replacing them
+    if os.path.exists(new_cpp_path) and os.path.exists(new_h_path):
+      # Check if the custom files exist
+      if os.path.exists(custom_new_cpp) and os.path.exists(custom_new_h):
+        print(f"Replacing {new_cpp_path} with custom new.cpp")
+        shutil.copy(custom_new_cpp, new_cpp_path)
+        print(f"Replacing {new_h_path} with custom new.h")
+        shutil.copy(custom_new_h, new_h_path)
+      else:
+        print("Custom new.cpp and new.h files not found.")
     else:
-      print("Custom new.cpp and new.h files not found!")
+      print(f"Original {new_cpp_path} or {new_h_path} not found, skipping replacement.")
 
 patch_new_lib()
